Får ikke skjult feltet du trykka på for  komme til rediger visningskolonner i en visning #1447
