+++ SharePointFramework/PortfolioWebParts/src/components/EditViewColumnsPanel/getItemStyle.tsx
@@ -1,0 +1,9 @@
+import {
+  DraggingStyle,
+  NotDraggingStyle
+} from 'react-beautiful-dnd'
+
+export const getItemStyle = (_isDragging: boolean, draggableStyle: DraggingStyle | NotDraggingStyle) => ({
+  userSelect: 'none',
+  ...draggableStyle
+})
