+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioOverview/PortfolioOverview.tsx
@@ -1,8 +1,8 @@
 import { WebPartContext } from '@microsoft/sp-webpart-base'
-import React, { FC } from 'react'
 import { EditViewColumnsPanel } from '../EditViewColumnsPanel'
+import React, { FC } from 'react'
 import { List } from '../List'
-import { ColumnContextMenu } from './ColumnContextMenu/ColumnContextMenu'
+import { ColumnContextMenu } from './ColumnContextMenu'
 import { ColumnFormPanel } from './ColumnFormPanel'
 import styles from './PortfolioOverview.module.scss'
 import { ViewFormPanel } from './ViewFormPanel'
