+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioOverview/useEditViewColumnsPanel.ts
@@ -2,7 +2,7 @@ import strings from 'PortfolioWebPartsStrings'
 import _ from 'lodash'
 import { ProjectColumn } from 'pp365-shared-library'
 import { useMemo } from 'react'
-import { IEditViewColumnsPanelProps } from '../EditViewColumnsPanel/types'
+import { EditViewColumnsPanelSortMode, IEditViewColumnsPanelProps } from '../EditViewColumnsPanel/types'
 import { IPortfolioOverviewContext } from './context'
 import { TOGGLE_EDIT_VIEW_COLUMNS_PANEL } from './reducer'
 
@@ -85,6 +85,6 @@ export function useEditViewColumnsPanel(
     },
     helpText: strings.PortfolioOverviewShowEditViewColumnsPanelHelpText,
     customColumnOrder: context.state.currentView?.columnOrder,
-    sortMode: 'customSelectedOnTop'
+    sortMode: EditViewColumnsPanelSortMode.CustomSelectedOnTop
   } as IEditViewColumnsPanelProps
 }
